Ex 4.15-4.23.

+++ part4/Blog List/controllers/blogsRouter.js
@@ -1,26 +1,55 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blogs')
+const User = require('../models/user')
+const { useExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+      .find({}).populate('user', { username: 1, name: 1})
 
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
-
-    if (!blog.title || !blog.url) {
+blogsRouter.post('/', useExtractor, async (request, response) => {
+    const body = request.body
+    const user = request.user
+    
+    if (!body.title || !body.url) {
         return response.status(400).json({ error: 'title or url missing' })
     }
 
+    const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes,
+        user: user._id
+    })
+
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', useExtractor, async (request, response) => {
+    const user = request.user
+    
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(404).json({error: 'blog not found'})
+    }
+    
+    if (blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
+    } 
+    
+    if (blog.user.toString() !== user.id.toString()){
+        return response.status(403).json({error: 'a blog can be deleted only by the user who added it'})
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
